+++ lib/mongoid/userstamp/mixins/model_mixin.rb
@@ -3,7 +3,7 @@
 module Mongoid
 module Userstamp
 
-  module Model
+  module ModelMixin
 
     extend ActiveSupport::Concern
 
@@ -12,7 +12,7 @@ module Userstamp
       belongs_to mongoid_userstamp_config.created_name, class_name: mongoid_userstamp_config.user_model, inverse_of: nil
       belongs_to mongoid_userstamp_config.updated_name, class_name: mongoid_userstamp_config.user_model, inverse_of: nil
 
-      before_create :set_created_by
+      before_validation :set_created_by
       before_save :set_updated_by
 
       protected
@@ -20,7 +20,9 @@ module Userstamp
       def set_created_by
         current_user = Mongoid::Userstamp.current_user(self.class.mongoid_userstamp_config.user_model)
         return if current_user.blank? || self.send(self.class.mongoid_userstamp_config.created_name)
-        self.send("#{self.class.mongoid_userstamp_config.created_name}=", current_user)
+        if self.send("#{self.class.mongoid_userstamp_config.created_name}").blank? && self.new_record?
+          self.send("#{self.class.mongoid_userstamp_config.created_name}=", current_user)
+        end
       end
 
       def set_updated_by