+++ spec/support/user.rb
@@ -1,7 +1,7 @@
 # -*- encoding : utf-8 -*-
 class User
   include Mongoid::Document
-  include Mongoid::Userstamp::User
+  include Mongoid::Userstamp::UserMixin
 
   field :name
 end